+++ src/main/java/com/contentful/java/cma/ModuleEntries.java
@@ -56,7 +56,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final String entryId = getResourceIdOrThrow(entry, "entry");
     final String spaceId = getSpaceIdOrThrow(entry, "entry");
 
-    return service.archive(spaceId, entryId, new Byte[0]).toBlocking().first();
+    return service.archive(spaceId, entryId).blockingFirst();
   }
 
   /**
@@ -85,9 +85,9 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
 
     try {
       if (entryId == null) {
-        return service.create(spaceId, contentTypeId, entry).toBlocking().first();
+        return service.create(spaceId, contentTypeId, entry).blockingFirst();
       } else {
-        return service.create(spaceId, contentTypeId, entryId, entry).toBlocking().first();
+        return service.create(spaceId, contentTypeId, entryId, entry).blockingFirst();
       }
     } finally {
       entry.setSystem(sys);
@@ -99,15 +99,15 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
    *
    * @param spaceId Space ID
    * @param entryId Entry ID
-   * @return String representing the success (203) of the action
+   * @return Integer representing the success (204) of the action
    * @throws IllegalArgumentException if spaceId is null.
    * @throws IllegalArgumentException if entry is null.
    */
-  public String delete(String spaceId, String entryId) {
+  public Integer delete(String spaceId, String entryId) {
     assertNotNull(spaceId, "spaceId");
     assertNotNull(entryId, "entryId");
 
-    return service.delete(spaceId, entryId).toBlocking().first();
+    return service.delete(spaceId, entryId).blockingFirst().code();
   }
 
   /**
@@ -134,7 +134,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
   public CMAArray<CMAEntry> fetchAll(String spaceId, Map<String, String> query) {
     assertNotNull(spaceId, "spaceId");
     DefaultQueryParameter.putIfNotSet(query, DefaultQueryParameter.FETCH);
-    return service.fetchAll(spaceId, query).toBlocking().first();
+    return service.fetchAll(spaceId, query).blockingFirst();
   }
 
   /**
@@ -147,7 +147,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
   public CMAEntry fetchOne(String spaceId, String entryId) {
     assertNotNull(spaceId, "spaceId");
     assertNotNull(entryId, "entryId");
-    return service.fetchOne(spaceId, entryId).toBlocking().first();
+    return service.fetchOne(spaceId, entryId).blockingFirst();
   }
 
   /**
@@ -164,8 +164,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final String entryId = getResourceIdOrThrow(entry, "entry");
     final String spaceId = getSpaceIdOrThrow(entry, "entry");
 
-    return service.publish(entry.getSystem().getVersion(), spaceId, entryId,
-        new Byte[0]).toBlocking().first();
+    return service.publish(entry.getSystem().getVersion(), spaceId, entryId).blockingFirst();
   }
 
   /**
@@ -182,7 +181,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final String entryId = getResourceIdOrThrow(entry, "entry");
     final String spaceId = getSpaceIdOrThrow(entry, "entry");
 
-    return service.unArchive(spaceId, entryId).toBlocking().first();
+    return service.unArchive(spaceId, entryId).blockingFirst();
   }
 
   /**
@@ -199,7 +198,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final String entryId = getResourceIdOrThrow(entry, "entry");
     final String spaceId = getSpaceIdOrThrow(entry, "entry");
 
-    return service.entriesUnPublish(spaceId, entryId).toBlocking().first();
+    return service.unPublish(spaceId, entryId).blockingFirst();
   }
 
   /**
@@ -221,7 +220,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final CMASystem system = entry.getSystem();
     entry.setSystem(null);
     try {
-      return service.update(version, spaceId, entryId, entry).toBlocking().first();
+      return service.update(version, spaceId, entryId, entry).blockingFirst();
     } finally {
       entry.setSystem(system);
     }
@@ -242,7 +241,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final String entryId = getResourceIdOrThrow(entry, "entry");
     final String spaceId = getSpaceIdOrThrow(entry, "entry");
 
-    return service.fetchAllSnapshots(spaceId, entryId).toBlocking().first();
+    return service.fetchAllSnapshots(spaceId, entryId).blockingFirst();
   }
 
   /**
@@ -264,7 +263,7 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
     final String entryId = getResourceIdOrThrow(entry, "entry");
     final String spaceId = getSpaceIdOrThrow(entry, "entry");
 
-    return service.fetchOneSnapshot(spaceId, entryId, snapshotId).toBlocking().first();
+    return service.fetchOneSnapshot(spaceId, entryId, snapshotId).blockingFirst();
   }
 
   /**
@@ -328,10 +327,11 @@ public final class ModuleEntries extends AbsModule<ServiceEntries> {
      * @throws IllegalArgumentException if spaceId is null.
      * @throws IllegalArgumentException if entry is null.
      */
-    public CMACallback<String> delete(final String spaceId, final String entryId,
-                                      CMACallback<String> callback) {
-      return defer(new RxExtensions.DefFunc<String>() {
-        @Override String method() {
+    public CMACallback<Integer> delete(final String spaceId,
+                                       final String entryId,
+                                       CMACallback<Integer> callback) {
+      return defer(new RxExtensions.DefFunc<Integer>() {
+        @Override Integer method() {
           return ModuleEntries.this.delete(spaceId, entryId);
         }
       }, callback);