+++ src/test/kotlin/com/contentful/java/cma/ContentTypeTests.kt
@@ -148,8 +148,8 @@ class ContentTypeTests : BaseTest() {
     }
 
     @test fun testDelete() {
-        val requestBody = "203"
-        server!!.enqueue(MockResponse().setResponseCode(200).setBody(requestBody))
+        val requestBody = ""
+        server!!.enqueue(MockResponse().setResponseCode(204).setBody(requestBody))
 
         assertTestCallback(client!!.contentTypes().async().delete(
                 "spaceid", "contenttypeid", TestCallback()) as TestCallback)
@@ -218,7 +218,7 @@ class ContentTypeTests : BaseTest() {
 
         // Request
         val request = server!!.takeRequest()
-        val url = HttpUrl.parse(server!!.url(request.path).toString())
+        val url = HttpUrl.parse(server!!.url(request.path).toString())!!
         assertEquals("1", url.queryParameter("skip"))
         assertEquals("2", url.queryParameter("limit"))
         assertEquals("bar", url.queryParameter("foo"))