+++ CHANGELOG.md
@@ -15,6 +15,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add: [Snapshots](https://www.contentful.com/developers/docs/references/content-management-api/#/reference/snapshots/entry-snapshots-collection).
 - Add: Query parameter for all collection endpoints. (skip and limit your responses)
 - Add: Ability to track rate limit headers.
+- Fix: Return HTTP Code of no content methods.
+- Fix: Assets: Do not send `url` and `details` if `uploadUrl` or `uploadFrom` is changed.
+- Polishing: Update dependencies
+  - retrofit 2.3.0 (was 2.0.1)
+  - commonsio 2.5 (was 2.4)
+  - junit: 4.12 (was 4.11)
+  - okhttp: 3.8.1 (was 3.2.0)
+  - rxjava: 2.1.1 (was 1.0.13)
+  - mockito: 2.8.47 (was 1.10.8)
 
 ## Version [2.0.0] - 2017-04-19
 - New: [`Upload` module](https://www.contentful.com/developers/docs/references/content-management-api/#/reference/uploads).